+++ CotEditor/Sources/FindPanelTextView.swift
@@ -9,7 +9,7 @@
  
  ------------------------------------------------------------------------------
  
- © 2015-2016 1024jp
+ © 2015-2017 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -104,6 +104,13 @@ final class FindPanelTextView: NSTextView {
     override func insertTab(_ sender: Any?) {
         
         self.window?.makeFirstResponder(self.nextKeyView)
+    }
+    
+    
+    /// jump to the previous responder with tab key (standard NSTextField behavior)
+    override func insertBacktab(_ sender: Any?) {
+        
+        self.window?.makeFirstResponder(self.previousKeyView)
     }