+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 ### Fixes
 
 - Fix an issue where the find panel was over expanded when performing “Find All” with a long find string.
+- Fix an issue where the find panel didn't select previous field with Shift + Tab keys