+++ CHANGELOG.md
@@ -37,9 +37,15 @@ develop
 3.2.7
 --------------------------
 
+### Improvements
+
+- Spread background drawing over paper width on printing.
+
+
 ### Fixes
 
 - Fix an issue where the find panel was over expanded when performing “Find All” with a long find string.
+- Fix an issue where the find panel didn't select previous field with Shift + Tab keys