+++ CHANGELOG.md
@@ -7,6 +7,7 @@ develop
 
 ### New Features
 
+- Open document in the existing Untitled window that was created automatically on an open/reopen event, if exists.
 - Add new actions “Half-width to Full-width” and “Full-width to Half-width” to Text > Transformations menu.
 - Add new setting option “Reopen windows from last session on launch” in General pane.
 - Add new setting option “including whitespace-only lines” for “trim trailing whitespace” action in General pane.
@@ -15,7 +16,6 @@ develop
 ### Improvements
 
 - Drop support for __OS X 10.10 Yosemite__.
-- Open document in the existing Untitled window that was created automatically on an open/reopen event, if exists.
 - Add clear button to input fields in the find panel.
 - Scroll console view after getting a new message to make it visible.
 - Swap position of “View” with “Format” menu to conform to Apple's HIG.
@@ -28,7 +28,6 @@ develop
 - Some minor UI improvements and fixes.
 - Update Japanese localization to conform with the modern macOS localization rules.
 - Update inner source code to Swift 4.
-- Update build environment to Xcode 9.2 (SDK macOS 10.13.2).
 - [non-AppStore ver.] Now, the application updater (Sparkle) can download and update CotEditor automatically, as like before CotEditor was Sandboxed.
     - This feature can actually be used first updating CotEditor 3.3.0 to CotEditor 3.3.1 or later.