+++ CHANGELOG.md
@@ -19,6 +19,7 @@ develop
 - Add clear button to input fields in the find panel.
 - Scroll console view after getting a new message to make it visible.
 - Display sharing window within the target document window.
+- Remove the feature shares document content text from File menu (Sharing the document file still exists.).
 - Swap position of “View” with “Format” menu to conform to Apple's HIG.
 - Display full encoding name in the status bar instead of the IANA charset name.
 - Add tooltips to Unicode normalization forms in Text menu.