+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Change Log
 
 - Fix an issue where the find panel was over expanded when performing “Find All” with a long find string.
 - Fix an issue where the file size in the status bar was not updated after saving.
-- Fix an issue where the find panel didn't select previous field with Shift + Tab keys
+- Fix an issue where the find panel didn't select previous field with Shift + Tab keys.
 - Fix an issue where the application freezed by opening the File Mapping Confilicts list when filename conflict exists.